% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DB.R
\name{add_records}
\alias{add_records}
\title{Add records from tibble or data frame to database}
\usage{
add_records(conn, table, data, dry_run = FALSE, ...)
}
\arguments{
\item{conn}{DB connection object.}

\item{table}{Table name.}

\item{data}{Tibble object with the records to be inserted in the database.}

\item{dry_run}{Boolean flag to return the query without running it.}

\item{...}{Optional parameters, including a boolean flag to hide status
messages (\code{quiet}).}
}
\description{
Add records from tibble or data frame to database
}
\keyword{internal}
