% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_model}
\alias{age_model}
\title{\code{age_model}}
\format{
A data frame with 10131 rows and 13 variables:
\describe{
\item{ID_SAMPLE}{PENDING}
\item{est_age_provided}{PENDING}
\item{est_age_original}{PENDING}
\item{est_age_bacon_intcal13}{PENDING}
\item{est_age_bacon_intcal13_max}{PENDING}
\item{est_age_bacon_intcal13_min}{PENDING}
\item{est_age_bacon_intcal20_mean}{PENDING}
\item{est_age_bacon_intcal20_median}{PENDING}
\item{est_age_bacon_intcal20_uncert_5}{PENDING}
\item{est_age_bacon_intcal20_uncert_95}{PENDING}
\item{est_age_bacon_intcal20_uncert_25}{PENDING}
\item{est_age_bacon_intcal20_uncert_75}{PENDING}
\item{comment}{PENDING}
}
}
\usage{
age_model
}
\description{
A tibble with records for age_model.
}
\keyword{datasets}
