% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basin_size}
\alias{basin_size}
\title{\code{basin_size}}
\format{
A data frame with 7 rows and 2 variables:
\describe{
\item{ID_BASIN_SIZE}{PENDING}
\item{basin_desc}{PENDING}
}
}
\usage{
basin_size
}
\description{
A tibble with records for basin_size.
}
\keyword{datasets}
