% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catch_size}
\alias{catch_size}
\title{\code{catch_size}}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{ID_CATCH_SIZE}{PENDING}
\item{catch_size}{PENDING}
}
}
\usage{
catch_size
}
\description{
A tibble with records for catch_size.
}
\keyword{datasets}
