% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_info}
\alias{date_info}
\title{\code{date_info}}
\format{
A data frame with 916 rows and 11 variables:
\describe{
\item{ID_DATE_INFO}{PENDING}
\item{ID_ENTITY}{PENDING}
\item{avg_depth}{PENDING}
\item{thickness}{PENDING}
\item{lab_num}{PENDING}
\item{ID_MAT_DATED}{PENDING}
\item{dated_age}{PENDING}
\item{error_positive}{PENDING}
\item{error_negative}{PENDING}
\item{ID_DATE_TYPE}{PENDING}
\item{date_comments}{PENDING}
}
}
\usage{
date_info
}
\description{
A tibble with records for date_info.
}
\keyword{datasets}
