% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_type}
\alias{date_type}
\title{\code{date_type}}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{ID_DATE_TYPE}{PENDING}
\item{date_type}{PENDING}
}
}
\usage{
date_type
}
\description{
A tibble with records for date_type.
}
\keyword{datasets}
