% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embsecbio.R
\name{embsecbio}
\alias{embsecbio}
\title{EMBSeCBIO Database structure}
\usage{
embsecbio()
}
\value{
EMBSeCBIO structure (including tables and attributes).
}
\description{
EMBSeCBIO Database structure
}
\keyword{internal}
