% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entity}
\alias{entity}
\title{\code{entity}}
\format{
A data frame with 1811 rows and 10 variables:
\describe{
\item{ID_ENTITY}{PENDING}
\item{ID_SITE}{PENDING}
\item{entity_name}{PENDING}
\item{latitude}{PENDING}
\item{longitude}{PENDING}
\item{elevation}{PENDING}
\item{entity_type}{PENDING}
\item{mod_or_0ka_class}{PENDING}
\item{comments}{PENDING}
\item{source}{PENDING}
}
}
\usage{
entity
}
\description{
A tibble with records for entity.
}
\keyword{datasets}
