% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entity_pub}
\alias{entity_pub}
\title{\code{entity_pub}}
\format{
A data frame with 645 rows and 3 variables:
\describe{
\item{ID_ENTITY_PUB}{PENDING}
\item{ID_ENTITY}{PENDING}
\item{ID_PUB}{PENDING}
}
}
\usage{
entity_pub
}
\description{
A tibble with records for entity_pub.
}
\keyword{datasets}
