% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entity_type}
\alias{entity_type}
\title{\code{entity_type}}
\format{
A data frame with 25 rows and 4 variables:
\describe{
\item{ID_ENTITY_TYPE}{PENDING}
\item{entity_type}{PENDING}
\item{mod_or_0ka_class}{PENDING}
\item{comments}{PENDING}
}
}
\usage{
entity_type
}
\description{
A tibble with records for entity_type.
}
\keyword{datasets}
