% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-entity.R
\name{get_entities}
\alias{get_entities}
\alias{get_entities.site}
\alias{get_entities.default}
\alias{get_entities.tbl_df}
\title{Get entities linked to a pollen site}
\usage{
get_entities(.site_tbl, ...)

\method{get_entities}{site}(.site_tbl, ...)

\method{get_entities}{default}(.site_tbl, ...)

\method{get_entities}{tbl_df}(.site_tbl, ...)
}
\arguments{
\item{.site_tbl}{\code{tibble} object.}

\item{...}{Optional parameters.}
}
\description{
Get entities linked to a pollen site
}
\examples{
\dontrun{
`\%>\%` <- embsecbio::`\%>\%`
ents <- embsecbio::site \%>\%
  dplyr::slice(1:4) \%>\%
  embsecbio::get_entities()
}
}
