% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DB.R
\name{get_quiet}
\alias{get_quiet}
\title{Extract parameter named quiet}
\usage{
get_quiet(...)
}
\arguments{
\item{...}{Optional parameters.}
}
\value{
Value for \code{quiet}.
}
\description{
Extract parameter named quiet from a list of optional parameters, if not
found, then return \code{FALSE} as the default value for \code{quiet}.
}
\keyword{internal}
