% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DB.R
\name{insert}
\alias{insert}
\alias{insert.age_model}
\alias{insert.date_info}
\alias{insert.entity}
\alias{insert.entity_pub}
\alias{insert.pollen_data}
\alias{insert.pub}
\alias{insert.sample}
\alias{insert.site}
\title{Insert records in the DB}
\usage{
insert(data, ...)

\method{insert}{age_model}(data, conn, ...)

\method{insert}{date_info}(data, conn, ...)

\method{insert}{entity}(data, conn, ...)

\method{insert}{entity_pub}(data, conn, ...)

\method{insert}{pollen_data}(data, conn, ...)

\method{insert}{pub}(data, conn, ...)

\method{insert}{sample}(data, conn, ...)

\method{insert}{site}(data, conn, ...)
}
\arguments{
\item{data}{Data to be inserted, must have column names and records across
the rows. It can be generated using
\code{\link[tibble:tibble]{tibble::tibble}}.}

\item{...}{Optional parameters, including connection object (\code{conn})
and boolean flag to hide status messages (\code{quiet}).}

\item{conn}{DB connection object.}
}
\value{
ID of the inserted record(s).
}
\description{
Insert records in the DB
}
