% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-workbook.R
\name{inspect}
\alias{inspect}
\title{Match input data types with the expected ones}
\usage{
inspect(x, tb = class(x)[1], coerce = FALSE, default = -999999)
}
\arguments{
\item{x}{Tibble with the input data.}

\item{tb}{String with the table name.}

\item{coerce}{Boolean flag to indicate if columns should be coerced to the
right type.}

\item{default}{Numeric value to be used when coercing non-numeric values.}
}
\value{
Invisibly return \code{x}
}
\description{
Match input data types with the expected ones
}
\keyword{internal}
