% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-workbook.R
\name{log_warnings}
\alias{log_warnings}
\title{Log records with warnings}
\usage{
log_warnings(data, file, append = TRUE, sep = ",")
}
\arguments{
\item{data}{Tibble with the record data.}

\item{file}{Filename where logs will be stored.}

\item{append}{if \code{TRUE} the data \code{x} are appended to the
    connection.}

\item{sep}{a string used to separate columns.  Using \code{sep = "\t"}
    gives tab delimited output; default is \code{" "}.}
}
\description{
Log records with warnings
}
\keyword{internal}
