% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-workbook.R
\name{map_string}
\alias{map_string}
\title{Map string to reference list}
\usage{
map_string(str, ref = embsecbio()$tables)
}
\arguments{
\item{str}{String to match.}

\item{ref}{Vector with reference strings.}
}
\value{
Matching string (if any).
}
\description{
Map string to reference list
}
\keyword{internal}
