% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_dated}
\alias{mat_dated}
\title{\code{mat_dated}}
\format{
A data frame with 26 rows and 3 variables:
\describe{
\item{ID_MAT_DATED}{PENDING}
\item{ID_MAT_DATED_HIGH}{PENDING}
\item{mat_dated}{PENDING}
}
}
\usage{
mat_dated
}
\description{
A tibble with records for mat_dated.
}
\keyword{datasets}
