% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{msg}
\alias{msg}
\title{Display message inside a box}
\usage{
msg(str, quiet = FALSE, limit = 80, nl = TRUE)
}
\arguments{
\item{str}{String to display.}

\item{quiet}{Boolean flag to hide messages.}

\item{limit}{Maximum number of characters, output width.}

\item{nl}{Boolean to indicate whether  or not to insert a new line after
the message}
}
\description{
Display message inside a box
}
\examples{
embsecbio:::msg("A")
embsecbio:::msg("AB")
embsecbio:::msg("ABC")
embsecbio:::msg("ABCD")
embsecbio:::msg("ABCD", limit = 30)
embsecbio:::msg("ABCD", quiet = TRUE)
embsecbio:::msg("ABCD", nl = FALSE)

}
\keyword{internal}
