% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DB.R
\name{na}
\alias{na}
\title{Convert \code{NA} and \code{NULL} values to \code{"NULL"}}
\usage{
na(x, quote = FALSE)
}
\arguments{
\item{x}{Column entry, can be numeric, logical, or character.}

\item{quote}{Boolean flag to indicate whether or not to add quotes to
\code{x}.}
}
\value{
\code{MySQL} \code{NULL} friendly value.
}
\description{
Convert \code{NA} and \code{NULL} values to \code{"NULL"}
}
\examples{
rpd:::na(NA)
rpd:::na(1)
rpd:::na("A", TRUE)
}
\keyword{internal}
