% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{plot_age_depth}
\alias{plot_age_depth}
\title{Plot Age-Depth}
\usage{
plot_age_depth(conn, entity_name, date_fill = "#F1C40F", quiet = TRUE, ...)
}
\arguments{
\item{conn}{DB connection object.}

\item{entity_name}{String with the name of the entity.}

\item{date_fill}{String with Hexadecimal colour for the dates.}

\item{quiet}{Boolean flag to hide status messages.}

\item{...}{Extra parameters (unused).}
}
\value{
\code{ggplot2} object.
}
\description{
Plot Age-Depth
}
