% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollen_data}
\alias{pollen_data}
\title{\code{pollen_data}}
\format{
A data frame with 324681 rows and 4 variables:
\describe{
\item{ID_SAMPLE_TAXA}{PENDING}
\item{ID_SAMPLE}{PENDING}
\item{taxon_clean}{PENDING}
\item{taxon_count}{PENDING}
}
}
\usage{
pollen_data
}
\description{
A tibble with records for pollen_data.
}
\keyword{datasets}
