% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pub}
\alias{pub}
\title{\code{pub}}
\format{
A data frame with 196 rows and 2 variables:
\describe{
\item{ID_PUB}{PENDING}
\item{citation}{PENDING}
}
}
\usage{
pub
}
\description{
A tibble with records for pub.
}
\keyword{datasets}
