% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embsecbio.R
\name{read_workbook}
\alias{read_workbook}
\title{Read workbook}
\usage{
read_workbook(workbook, wdir = NA, sheets = list_sheets(), skip = 1, ...)
}
\arguments{
\item{workbook}{Workbook filename.}

\item{wdir}{Path to the workbook.}

\item{sheets}{Sheet names, by default it takes the names listed in the
description.}

\item{skip}{Number of lines to skip.}

\item{...}{
  Arguments passed on to \code{\link[readxl:read_excel]{readxl::read_excel}}
  \describe{
    \item{\code{range}}{A cell range to read from, as described in \link[readxl]{cell-specification}.
Includes typical Excel ranges like "B3:D87", possibly including the sheet
name like "Budget!B2:G14", and more. Interpreted strictly, even if the
range forces the inclusion of leading or trailing empty rows or columns.
Takes precedence over \code{skip}, \code{n_max} and \code{sheet}.}
    \item{\code{col_names}}{\code{TRUE} to use the first row as column names, \code{FALSE} to get
default names, or a character vector giving a name for each column. If user
provides \code{col_types} as a vector, \code{col_names} can have one entry per
column, i.e. have the same length as \code{col_types}, or one entry per
unskipped column.}
    \item{\code{na}}{Character vector of strings to interpret as missing values. By
default, readxl treats blank cells as missing data.}
    \item{\code{trim_ws}}{Should leading and trailing whitespace be trimmed?}
    \item{\code{n_max}}{Maximum number of data rows to read. Trailing empty rows are
automatically skipped, so this is an upper bound on the number of rows in
the returned tibble. Ignored if \code{range} is given.}
    \item{\code{guess_max}}{Maximum number of data rows to use for guessing column
types.}
    \item{\code{progress}}{Display a progress spinner? By default, the spinner appears
only in an interactive session, outside the context of knitting a document,
and when the call is likely to run for several seconds or more. See
\code{\link[readxl:readxl_progress]{readxl_progress()}} for more details.}
    \item{\code{.name_repair}}{Handling of column names. By default, readxl ensures
column names are not empty and are unique. If the tibble package version is
recent enough, there is full support for \code{.name_repair} as documented in
\code{\link[tibble:tibble]{tibble::tibble()}}. If an older version of tibble is present, readxl falls
back to name repair in the style of tibble v1.4.2.}
  }}
}
\value{
List with tibbles for each sheet in the workbook.
}
\description{
Read workbook. This is expected to have at least three sheets with the
following names:
\describe{
\item{\code{Site Metadata}}{}
\item{\code{Entity metadata}}{}
\item{\code{SampleCharcoalChronology}}{}
\item{\code{Date Info}}{}
}
}
\keyword{internal}
