% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-workbooks.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace missing data}
\usage{
replace_missing(data)
}
\arguments{
\item{data}{Input data (column)}
}
\value{
Input data with the missing filters replaced by their corresponding
codes.
}
\description{
Replace missing data
}
\details{
\describe{
\item{\verb{not applicable} -> \code{NULL}}{for fields where we would not expect
an answer (e.g. water depth for terrestrial sites, lab number/material
dated for pollen correlations).}
\item{\verb{not recorded} -> \code{-999999}}{for fields where we/they have checked
and there is no information (information not collected, contributor has
lost data).}
\item{\verb{not known} -> \code{-777777}}{for fields where we don't know what the
answer is but there could be one and we should check at some stage}
}
}
\keyword{internal}
