% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample}
\alias{sample}
\title{\code{sample}}
\format{
A data frame with 11232 rows and 6 variables:
\describe{
\item{ID_SAMPLE}{PENDING}
\item{ID_ENTITY}{PENDING}
\item{sample_name}{PENDING}
\item{avg_depth}{PENDING}
\item{sample_type}{PENDING}
\item{count_type}{PENDING}
}
}
\usage{
sample
}
\description{
A tibble with records for sample.
}
\keyword{datasets}
