% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_type}
\alias{sample_type}
\title{\code{sample_type}}
\format{
A data frame with 49 rows and 2 variables:
\describe{
\item{ID_SAMPLE_TYPE}{PENDING}
\item{sample_type}{PENDING}
}
}
\usage{
sample_type
}
\description{
A tibble with records for sample_type.
}
\keyword{datasets}
