% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_type}
\alias{site_type}
\title{\code{site_type}}
\format{
A data frame with 34 rows and 3 variables:
\describe{
\item{ID_SITE_TYPE}{PENDING}
\item{ID_HIGHER}{PENDING}
\item{desc_site_type}{PENDING}
}
}
\usage{
site_type
}
\description{
A tibble with records for site_type.
}
\keyword{datasets}
