% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-workbook.R
\name{update_names}
\alias{update_names}
\title{Update column names}
\usage{
update_names(vars, sheet = NA)
}
\arguments{
\item{vars}{Vector of strings with the original names.}

\item{sheet}{String with workbook sheet name.}
}
\value{
Updated names.
}
\description{
Update column names from workbook sheets to match the names in the tables of
the database.
}
\keyword{internal}
