% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-DB.R
\name{update_records}
\alias{update_records}
\title{Update records from tibble or data frame to database}
\usage{
update_records(conn, table, data, dry_run = FALSE, PK = 1, ...)
}
\arguments{
\item{conn}{DB connection object.}

\item{table}{Table name.}

\item{data}{Tibble object with the records to be updated in the database.}

\item{dry_run}{Boolean flag to return the query without running it.}

\item{PK}{Numeric vector with the indices of the columns to use as primary
key.}

\item{...}{Optional parameters, including a boolean flag to hide status
messages (\code{quiet}).}
}
\description{
Update records from tibble or data frame to database
}
\keyword{internal}
