% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site}
\alias{site}
\title{\code{site}}
\format{
A data frame with 1463 rows and 8 variables:
\describe{
\item{ID_SITE}{PENDING}
\item{site_name}{PENDING}
\item{latitude}{PENDING}
\item{longitude}{PENDING}
\item{elevation}{PENDING}
\item{ID_SITE_TYPE}{PENDING}
\item{ID_BASIN_SIZE}{PENDING}
\item{ID_CATCH_SIZE}{PENDING}
}
}
\usage{
site
}
\description{
A tibble with records for site.
}
\keyword{datasets}
